feat: add putting lotto data from view into DB
