+++ src/main/java/lotto/WebUILottoApplication.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.db.dao.LottoDAO;
 import lotto.domain.Factory.LottoTicketsFactory;
 import lotto.domain.LottoTickets;
 import lotto.domain.Money;
@@ -31,17 +32,16 @@ public class WebUILottoApplication {
             // money, lottos('-'를 기준으로 한 숫자 문자열)
             get("/ticket", (req, res) -> {
                 Map<String, Object> model = new HashMap<>();
-
                 Money money = new Money(Integer.parseInt(req.queryParams("money")));
-                model.put("money", money.getMoney());
-
                 List<String> inputCustoms = Arrays.asList(req.queryParams("lottos").split("-"));
                 LottoTickets lottoTickets = LottoTicketsFactory.getInstance().create(money, inputCustoms);
 
+                model.put("money", money.getMoney());
                 model.put("lottos", lottoTickets.getLottoTickets());
-
                 model.put("amountOfCustom", inputCustoms.size());
                 model.put("amountOfAuto", (money.getMoney() / 1000) - inputCustoms.size());
+
+                LottoDAO.addLottoTicket(lottoTickets);
 
                 return render(model, "purchased_tickets.html");
             });
@@ -50,21 +50,16 @@ public class WebUILottoApplication {
             // money, lottos('-'를 기준으로 한 숫자 문자열)
             post("/ticket", (req, res) -> {
                 Map<String, Object> model = new HashMap<>();
-
                 Money money = new Money(Integer.parseInt(req.queryParams("money")));
-                model.put("money", money.getMoney());
-
                 List<String> inputCustoms = Arrays.asList(req.queryParams("lottos").split("-"));
                 LottoTickets lottoTickets = LottoTicketsFactory.getInstance().create(money, inputCustoms);
 
-                List<String> customs = new ArrayList<>();
-                for (String custom : customs) {
-
-                }
+                model.put("money", money.getMoney());
                 model.put("lottos", lottoTickets.getLottoTickets());
-
                 model.put("amountOfCustom", inputCustoms.size());
                 model.put("amountOfAuto", (money.getMoney() / 1000) - inputCustoms.size());
+
+                LottoDAO.addLottoTicket(lottoTickets);
 
                 return render(model, "purchased_tickets.html");
             });
