+++ src/main/java/lotto/db/dao/WinningLottoDAO.java
@@ -1,7 +1,6 @@
-package lotto.dao;
+package lotto.db.dao;
 
-import lotto.dto.LottoDTO;
-import lotto.dto.WinningLottoDTO;
+import lotto.db.dto.WinningLottoDTO;
 
 import java.sql.*;
 import java.util.Arrays;
@@ -9,17 +8,18 @@ import java.util.List;
 
 import static lotto.db.DBConnection.getConnection;
 
-public class LottoDAO {
+public class WinningLottoDAO {
     private static final Connection conn = getConnection();
 
-    public void addLottoTicket(LottoDTO lotto) throws SQLException {
+    public void addWinningLottoTicket(WinningLottoDTO winningLotto) throws SQLException {
         int autoInsertedKey = addLotto();
-        addLottoNumbers(autoInsertedKey, lotto.getNumbers());
+        addLottoNumbers(autoInsertedKey, winningLotto.getNumbers());
+        addWinningLotto(winningLotto, autoInsertedKey);
     }
 
     private int addLotto() throws SQLException {
-        // 일반로또의 로또 타입은 0
-        String query = "INSERT INTO lotto.lotto (type) VALUES (0)";
+        // 당첨로또의 로또 타입은 1
+        String query = "INSERT INTO lotto.lotto (type) VALUES (1)";
         PreparedStatement pstmt = conn.prepareStatement(query, Statement.RETURN_GENERATED_KEYS);
 
         pstmt.executeUpdate();
@@ -41,35 +41,46 @@ public class LottoDAO {
         pstmt.executeBatch();
     }
 
-    public LottoDTO findByLottoId(String lottoId) throws SQLException {
-        String query = "SELECT l.id, l.type, GROUP_CONCAT(ln.number SEPARATOR  ',') as numbers " +
-                "FROM lotto.lotto as l " +
-                "JOIN lotto.lottonumber as ln ON l.id = ln.lotto_id " +
-                "WHERE l.id = ? " +
-                "GROUP BY l.id";
+    private void addWinningLotto(WinningLottoDTO winningLotto, int autoInsertedKey) throws SQLException {
+        String query = "INSERT INTO lotto.winninglotto (lotto_id, bonusBall) VALUES (?, ?)";
         PreparedStatement pstmt = conn.prepareStatement(query);
-        pstmt.setString(1, lottoId);
-        ResultSet rs = pstmt.executeQuery();
 
-        return getLottoDTO(rs);
+        pstmt.setInt(1, autoInsertedKey);
+        pstmt.setInt(2, winningLotto.getBonusBall());
+
+        pstmt.executeUpdate();
     }
 
-    private LottoDTO getLottoDTO(ResultSet rs) throws SQLException {
+    public WinningLottoDTO findByWinningLottoId(String winningLottoId) throws SQLException {
+        String query = "SELECT l.id, l.type, GROUP_CONCAT(ln.number SEPARATOR ',') as numbers, w.bonusBall " +
+                "FROM lotto.lotto as l " +
+                "JOIN lotto.winninglotto as w ON w.lotto_id = l.id " +
+                "JOIN lotto.lottonumber as ln ON l.id = ln.lotto_id " +
+                "WHERE w.id = ? " +
+                "GROUP BY l.id;";
+        PreparedStatement pstmt = conn.prepareStatement(query);
+        pstmt.setString(1, winningLottoId);
+        ResultSet rs = pstmt.executeQuery();
+
+        return getWinningLottoDTO(rs);
+    }
+
+    private WinningLottoDTO getWinningLottoDTO(ResultSet rs) throws SQLException {
         if (!rs.next()) {
             return null;
         }
 
         try {
             int id = rs.getInt("id");
+            int bonusBall = rs.getInt("bonusBall");
             List<String> numbers = Arrays.asList(rs.getString("numbers").split(","));
 
             // TODO 나중에 빌더 패턴으로 바꿔보기
-            return new LottoDTO(id,
+            return new WinningLottoDTO(id, bonusBall,
                     Integer.parseInt(numbers.get(0)), Integer.parseInt(numbers.get(1)), Integer.parseInt(numbers.get(2)),
                     Integer.parseInt(numbers.get(3)), Integer.parseInt(numbers.get(4)), Integer.parseInt(numbers.get(5)));
         } catch (Exception e) {
             return null;
         }
     }
-
 }
