+++ src/main/java/lotto/db/dto/LottoDTO.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.db.dto;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -13,16 +13,6 @@ public class LottoDTO {
     private int num4;
     private int num5;
     private int num6;
-
-    public LottoDTO(int lottoId, int num1, int num2, int num3, int num4, int num5, int num6) {
-        this.lottoId = lottoId;
-        this.num1 = num1;
-        this.num2 = num2;
-        this.num3 = num3;
-        this.num4 = num4;
-        this.num5 = num5;
-        this.num6 = num6;
-    }
 
     public LottoDTO(int num1, int num2, int num3, int num4, int num5, int num6) {
         this.num1 = num1;
